bot/requirements.txt
python-telegram-bot==20.8
torch==2.2.0
transformers==4.36.2 
